Implement initialized static variable definitions for scalar integer types (task #30).
- parser/expr.c: fix unsuffixed integer literals to pick the smallest type
  that holds the value (LITERAL_INT / LITERAL_LONG / LITERAL_LONG_LONG)
  instead of always truncating to LITERAL_INT; large octal/hex constants
  such as 01234567076543210 no longer lose their upper bits
- semantic/const_convert.c: map TYPE_UINT → TAC_STATIC_INIT_U64 so the
  full 48-bit BESM-6 unsigned word is preserved
- semantic/const_convert_tests.cpp: update UIntFrom* expectations to U64
- backend/besm6/besm.h: widen Besm_DataItem.log_val to unsigned long long
- backend/besm6/emit_madlen.c: use %llo to print 48-bit LOG values
- backend/besm6/codegen.c: implement codegen_static_variable for the
  initialized case (BSS remains unchanged); add static_init_log_val helper
  that converts each integer TAC_STATIC_INIT_* to a raw 48-bit octal word
- backend/besm6/init_tests.cpp: enable five tests (VarIntInit, VarLongInit,
  VarShortInit, VarCharInit, VarUnsignedInit)

+++ backend/besm6/besm.h
@@ -129,7 +129,7 @@ struct Besm_Func {
 typedef enum {
     BESM_DATA_INT,    // DI_Int(int value)     -- ,INT, value
     BESM_DATA_REAL,   // DI_Real(double value) -- ,REAL, value
-    BESM_DATA_LOG,    // DI_Log(int value)     -- ,LOG, value (logical constant)
+    BESM_DATA_LOG,    // DI_Log(unsigned long long value) -- ,LOG, value (logical constant)
     BESM_DATA_BSS,    // DI_Bss(int words)     -- ,BSS, count (zero words)
     BESM_DATA_EQU,    // DI_Equ(int value)     -- ,EQU, value (symbolic constant)
     BESM_DATA_REF,    // DI_Ref(string name)   -- word-wide label reference
@@ -142,7 +142,7 @@ struct Besm_DataItem {
     union {
         int int_val;      // DI_Int
         double real_val;  // DI_Real
-        int log_val;      // DI_Log
+        unsigned long long log_val; // DI_Log
         int bss_words;    // DI_Bss
         int equ_val;      // DI_Equ
         char *ref_name;   // DI_Ref (heap-owned)
